2020-13
I had a spoiler for the second part. My existing solutions were too slow for the final one. I'd missed the way that this can incremement (and not even convinced I understand it). I tried using the Chinese Remainder Theorem, but got the wrong answers. Not sure where it went wrong.